Stats, movement & plants

+++ darwin_world/src/main/java/agh/oop/simulation/AnimalFateStatistics.java
@@ -12,4 +12,4 @@ public record AnimalFateStatistics(
         int numberOfChildren,
         int numberOfDescendants,
         long age,
-        int deathDay) {}
+        long deathDay) {}
